pyproject.toml
[tool.poetry]
name = "mcgill-gpt"
version = "0.1.0"
description = ""
authors = ["Adam Geenen <adamgeenen@gmail.com>"]
readme = "README.md"
packages = [{include = "mcgill_gpt"}]

[tool.poetry.dependencies]
python = "^3.9.12"
openai = "^0.27.8"
langchain = "^0.0.250"
streamlit = "^1.25.0"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
beautifulsoup4 = "^4.12.2"
unstructured = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
